+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 - Changed `AbstractAuthorizedHttpExchangeClient` to an interface and renamed it to `AuthorizedHttpExchangeClient`.
 - Added `ScopedResource.closeIfManaged(Consumer)` which can handle resource closing exceptions.
 - Added constructor with resource only to `ScopedResource` defaulting to a managed resource.
-- Fixed `HttpContentType.parse` methods when `charset` parameter is missing but other parameter is present. 
+- Fixed `HttpContentType.parse` methods when `charset` parameter is missing but other parameter is present.
+- Renamed `ExchangeClientBuilder.decorateWith` with builder class parameter to `ExchangeClientBuilder.decorateWithBuilder`.
+- Added `ExchangeClientBuilder.decorateWith` with `DecoratingExchangeClient` class as parameter.
+- Added `ApiClient.with(ExchangeClient)` static method to be used when the exchange client is managed by the caller.
 
 ---